fix(app_policy): manifest_json is not mandatory
In an app policy a `manifest_json` block is not actually mandatory.
It is not even clear if on the Zedcloud side the information inside
the `manifest_json` block is actually used since the referenced app
will contain it's own "manifest".

Additionally if we try to include a minimal `manifest_json` block in
the TF config it will cause an error on the Zedcloud side since
`app_type` and `deployment_type` will be set to empty strings which
are not valid values for those "enums".